Merge

+++ src/server/shared/Cryptography/ARC4.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2008-2012 TrinityCore <http://www.trinitycore.org/>
+ * Copyright (C) 2008-2015 TrinityCore <http://www.trinitycore.org/>
  * Copyright (C) 2005-2009 MaNGOS <http://getmangos.com/>
  *
  * This program is free software; you can redistribute it and/or modify it
@@ -19,14 +19,14 @@
 #include "ARC4.h"
 #include <openssl/sha.h>
 
-ARC4::ARC4(uint8 len) : m_ctx()
+ARC4::ARC4(uint32 len) : m_ctx()
 {
     EVP_CIPHER_CTX_init(&m_ctx);
     EVP_EncryptInit_ex(&m_ctx, EVP_rc4(), NULL, NULL, NULL);
     EVP_CIPHER_CTX_set_key_length(&m_ctx, len);
 }
 
-ARC4::ARC4(uint8 *seed, uint8 len) : m_ctx()
+ARC4::ARC4(uint8 *seed, uint32 len) : m_ctx()
 {
     EVP_CIPHER_CTX_init(&m_ctx);
     EVP_EncryptInit_ex(&m_ctx, EVP_rc4(), NULL, NULL, NULL);
