+++ src/server/shared/Cryptography/ARC4.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2008-2012 TrinityCore <http://www.trinitycore.org/>
+ * Copyright (C) 2008-2015 TrinityCore <http://www.trinitycore.org/>
  * Copyright (C) 2005-2009 MaNGOS <http://getmangos.com/>
  *
  * This program is free software; you can redistribute it and/or modify it
@@ -19,14 +19,14 @@
 #ifndef _AUTH_SARC4_H
 #define _AUTH_SARC4_H
 
-#include "Define.h"
 #include <openssl/evp.h>
+#include "Define.h"
 
 class ARC4
 {
     public:
-        ARC4(uint8 len);
-        ARC4(uint8 *seed, uint8 len);
+        ARC4(uint32 len);
+        ARC4(uint8 *seed, uint32 len);
         ~ARC4();
         void Init(uint8 *seed);
         void UpdateData(int len, uint8 *data);
